refactor: create feature folder for domain-specific components
Organized feature-specific components into feature/ directory by domain:
- feature/auth/ - Authentication components
- feature/collect/ - Collection and search components
- feature/home/ - Home page components
- feature/project/ - Project-specific components
- feature/portfolio/ - Portfolio-specific components

File moves:
- auth/* → feature/auth/
  - Account.tsx, GoogleOneTab.tsx, LoginBox.tsx, SupabaseSessionSync.tsx
- collect/* → feature/collect/
  - CollectClient.tsx, constants.ts
- HomeProjectList.tsx → feature/home/HomeProjectList.tsx
- project/* → feature/project/
  - ProjectPage.tsx, staticParamsGenerator.ts
  - components/ProjectEditModal.tsx
  - components/ProjectGrid.tsx
  - components/ProjectMainContent.tsx
- contents/ProfileItem.tsx → feature/portfolio/ProfileItem.tsx

Updated imports in 15 files:
- Auth: layout.tsx, page.tsx, Header.tsx
- Collect: project/page.tsx, team/page.tsx
- Home: page.tsx
- Project: openProjectModal.tsx, ProjectSidebar.tsx, HomeProjectList.tsx,
  PortfolioHome.tsx, Team.tsx, ProjectPage.tsx, project detail pages
- Portfolio: TeamSidebar.tsx, PortfolioDetail.tsx

This creates clear separation of domain-specific logic and improves
component discoverability.

Verified with pnpm lint and pnpm build.
