+++ src/app/(box-layout)/team/Team.tsx
@@ -4,7 +4,7 @@ import { getTeamProjects } from '@/services/team/getTeamProjects.server';
 import { notFound } from 'next/navigation';
 import TeamSidebar from '@/app/components/sidebar/TeamSidebar';
 import ProfileIcon from '@/app/components/card/portfolio/components/ProfileIcon';
-import ProjectGrid from '@/app/components/project/components/ProjectGrid';
+import ProjectGrid from '@/app/components/feature/project/components/ProjectGrid';
 import SidebarContentLayout from '@/app/components/layout/sidebar/SidebarContentLayout';
 
 const Team = async ({ teamName }: { teamName: string }) => {
