+++ src/app/components/card/portfolio/PortfolioHome.tsx
@@ -4,7 +4,7 @@ import { Title } from '../../ui/text/text';
 import { PortfolioDetailProps } from '@/app/(box-layout)/portfolio/types';
 import { CardProps } from '@/app/components/card/project/ProjectCard';
 import ProfileEditButton from './ProfileEditButton';
-import ProjectGrid from '@/app/components/project/components/ProjectGrid';
+import ProjectGrid from '@/app/components/feature/project/components/ProjectGrid';
 
 interface PortfolioHomeProps {
   content: string;
