+++ src/app/components/feature/project/staticParamsGenerator.ts
